pyproject.toml
[project]
name = "parladach-backend"
version = "0.1.0"
description = "Backend FastAPI - Parladach"
requires-python = ">=3.11"
dependencies = [
  "fastapi>=0.110",
  "uvicorn[standard]>=0.27",
  "pydantic-settings>=2.2",
  "sqlalchemy>=2.0",
  "psycopg[binary]>=3.2",
  "alembic>=1.13",
  "pwdlib[argon2]>=0.3.0",
  "email-validator>=2.1.1",
  "PyJWT>=2.8.0",
]

[project.optional-dependencies]
dev = [
  "pytest>=8.0",
  "httpx>=0.27",
]

[tool.pytest.ini_options]
pythonpath = ["src"]
testpaths = ["tests"]

